Create Situated- and Separate- MessageEditor
